% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example_lapacke_dgels}
\alias{example_lapacke_dgels}
\title{Example: using lapacke_dgels}
\usage{
example_lapacke_dgels()
}
\description{
Solve linear systems using lapacke_dgels
}
\examples{
## Expected output:
example_lapacke_dgels()
}
