% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lapacker}
\alias{lapacker}
\alias{lapacker-package}
\title{lapacker}
\description{
Provide the 'lapacke.h' header file for R. 'LAPACKE' is a
collection of wrappers that provide a C interface to the FORTRAN 'LAPACK'
library. Currently only the double precision / double complex 'LAPACK'
routines are supported.
}
\author{
Yi Pan
}
